fix: change NMAMinorTraceChemistry volume from Float to Integer
Update volume field type to match source database schema (NM_Aquifer_Dev_DB).

Changes:
- Update model type in db/nma_legacy.py
- Add _safe_int() helper in transfer script
- Update tests to use integer values
- Add Alembic migration to alter column type

Closes #379

+++ db/nma_legacy.py
@@ -375,7 +375,7 @@ class NMAMinorTraceChemistry(Base):
     notes: Mapped[Optional[str]] = mapped_column(Text)
     analyses_agency: Mapped[Optional[str]] = mapped_column(String(100))
     uncertainty: Mapped[Optional[float]] = mapped_column(Float)
-    volume: Mapped[Optional[float]] = mapped_column(Float)
+    volume: Mapped[Optional[int]] = mapped_column(Integer)
     volume_unit: Mapped[Optional[str]] = mapped_column(String(20))
 
     # --- Relationships ---
